Implement enhanced backup configuration and management features
- Update .env.example to include new backup settings such as encryption key, alerting webhooks, and offsite backup options.
- Modify docker-compose.yml to remove Redis service and integrate new backup environment variables.
- Refactor create_backup.sh to use a stronger default password and improve backup process.
- Enhance backup.py to support file compression and accurate size reporting, along with improved error handling.
- Update backup scripts to ensure secure backup encryption and status logging.
- Adjust health_records.py to allow for a higher limit on record retrieval.
- Improve Admin and Records Vue components for better user experience and data handling.

+++ backend/app/models/backup.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, DateTime, Integer, Enum, Text, Boolean
+from sqlalchemy import Column, String, DateTime, Integer, Float, Enum, Text, Boolean
 from datetime import datetime, timezone
 import enum
 
@@ -24,7 +24,7 @@ class BackupHistory(Base):
     
     file_path = Column(String, nullable=True)
     file_size = Column(Integer, nullable=True)
-    size_mb = Column(Integer, nullable=True)  # Size in MB for display
+    size_mb = Column(Float, nullable=True)  # Size in MB for display
     checksum = Column(String, nullable=True)
     duration_seconds = Column(Integer, nullable=True)  # Duration in seconds
